[UPDATE] formatCountNumber, formatTime 함수 이동
- 두 힘수는 VideoItem에서만 쓰이는 함수가 아니기 때문에 글로벌 함수로 분리함
- 글로벌한 함수로 분리하면서 인자명도 일반적인 이름으로 변경함(createdTime->timestamp)
- JSDoc 추가함. 함수 정의할 때 description, param, return정도는 적어주기
